+++ webapp/client/apps/orchestration-cluster-webapp/src/modules/login/getCurrentCopyrightNoticeText.ts
