+++ webapp/client/apps/orchestration-cluster-webapp/src/pages/login.tsx
@@ -9,11 +9,11 @@
 import {Form, Field} from 'react-final-form';
 import {FORM_ERROR} from 'final-form';
 import {TextInput, PasswordInput, Column, Grid, Stack, InlineNotification, Button} from '@carbon/react';
-import {LoadingSpinner} from './LoadingSpinner';
+import {LoadingSpinner} from '../modules/login/components/LoadingSpinner';
 import {CamundaLogo} from '#/modules/components/CamundaLogo';
-import {getCurrentCopyrightNoticeText} from './getCurrentCopyrightNoticeText';
-import {Disclaimer} from './Disclaimer';
-import styles from './styles.module.scss';
+import {getCurrentCopyrightNoticeText} from '../modules/login/getCurrentCopyrightNoticeText';
+import {Disclaimer} from '../modules/login/components/Disclaimer';
+import styles from './login.module.scss';
 
 type FormValues = {
 	username: string;
