+++ webapp/client/apps/orchestration-cluster-webapp/src/routes/login.tsx
@@ -7,7 +7,7 @@
  */
 
 import {createFileRoute} from '@tanstack/react-router';
-import {Login} from '../pages/Login';
+import {Login} from '../pages/login';
 
 export const Route = createFileRoute('/login')({
 	component: Login,
